[multi-dut] add infrastructure to update dut port to ptf port index map
In the single DUT scenario, the DUT interface indeces exactly match with
the ptf port indeces. However, in multi-dut scneario, this is not always
true. Multiple DUTs are sharing the same ptf host, therefore their interface
indeces need to be translated to ptf port indeces according to the topology
definition.

In this change, a translate map is captured from topology definition. And
member method get_minigraph_facts() is added to SonicHost to help getting
the minigrap facts and then update ptf port indeces.

Changed test_lag_2 to use the new method to map to the correct ptf port.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>