[multi-dut] add infrastructure to update dut port to ptf port index map (#2460)
Description of PR
Summary:
Fixes # (issue)

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 Test case(new/improvement)
Approach
What is the motivation for this PR?
In the single DUT scenario, the DUT interface indeces exactly match with the ptf port indeces. However, in multi-dut scneario, this is not always true. Multiple DUTs are sharing the same ptf host, therefore their interface indeces need to be translated to ptf port indeces according to the topology definition.

How did you do it?
In this change, a translate map is captured from topology definition. And member method get_extended_minigraph_facts() is added to SonicHost to help getting the minigrap facts and then extend it with ptf port indices with mg_facts['minigraph_ptf_indeces'].

Changed test_lag_2 to use the new method to map to the correct ptf port.

Signed-off-by: Ying Xie ying.xie@microsoft.com

How did you verify/test it?
Tested the infrastructure with updated lag_2 test.