[warm-reboot test] Update warm-reboot test to support kvm image (#2392)
What is the motivation for this PR?
The motivation of this PR is to enable basic checks of the warm-reboot functionality using kvm images.

How did you do it?
As the data plane will go down in kvm image during warm-reboot. I changed the test logic to mainly focusing on the control plane behavior and relax data plane constraints.

A mark is added in the testcase to indicate it support kvm image and to filter out testcases not compatible for kvm test.