[SonicHost] Ignore non-zero return code from 'supervisorctl status' in critical_process_status() (#2474)
Ignore non-zero return code from `supervisorctl status` command when called from `SonicHost.critical_process_status()`

As part of migrating all Python code in SONiC to Python 3, we must upgrade supervisor to version >= 4.0.0, as this is the first version which supports Python 3. However, as of version 4.0.0, the return codes from `supervisorctl status` have changed, such that if _any_ process is not in a `RUNNING` state--even if it is expected (expected exit or even stopped by user)--it will return `3`. Thus, we need to ignore the return code from `supervisorctl status`.

- Example of the issue: https://sonic-jenkins.westus2.cloudapp.azure.com/job/vs/job/buildimage-vs-image-pr/6854/console

- Discussion of supervisorctl behavior change: https://github.com/Supervisor/supervisor/issues/1223