Reusable fixture for selective enable and disable of service autorestart (#2458)
Summary: Add a callable feature for autorestart enable/disable

Type of change: Test case(new/improvement)

Motivation: When a service is stopped as part of a standalone test, the service restarts due to feature autorestart is enabled. This error is not seen as part of nightly jobs as test_pretest and test_posttest make sure that the service does not restart as part of the testcase after stopping it.

What changed: Moved the feature autorestart enable disable code to confttest. However, this time the logic is callable and every test can use this to enable and disable feature autorestart.

Verification: Tested test_qos_sai testcase and the disabled service autorestart before stopping them.