Move SONiC node type checking methods from DutHosts to SonicHost (#2543)
For multi-DUT setup, we need a convenient way to check the node type. Originally
we have methods is_supervisor_node and is_frontend_node defined in class DutHosts. 
It turns out that the better place to define these methods is in the SonicHost class. 
This change refactored the DutHosts class and moved these two methods to the 
SonicHost class. The affected script test_bgp_fact.py is also updated.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>