[Kubernetes Testing] Add Configured DNS Server for Physical DUT Testbed (#2579)
Added optional dns_server_ip spec to host_vars ansible var files for testbed server. When dns_server_ip is defined, the KVMs will be configured with this DNS server IP. When dns_server_ip is not defined, the proxy server is likely configured and will take care of DNS when necessary packages are installed as part of HA master setup process.