Fix bug in test_bgpmon
The ptfadapter is a module level fixture, which means it's not cleared
between test_bgpmon and test_bgpmon_no_resolve_via_default. As a result,
the TCP sync packet captured in test_bgpmon might be counted in
test_bgpmon_no_resolve_via_default and cause test failure.
This commit add a flush before beginning counting packet to address the
issue.

Signed-off-by: bingwang <bingwang@microsoft.com>