Improve the pytest port_toggle library
PR https://github.com/Azure/sonic-mgmt/pull/2548 fixed a port_toggle
issue in the ansible playbook. The pytest port_toggle library also has the
similar issue. Description of the issue in PR#2548: "After taking all ports
down, it only verify that all ports in linux side is down, but it does not
check that all ports in DB is really down.  The fix is add a check of port
status returned by "show_interface" module.

The second issue of this library is that it depends on a function defined
in platform_tests/link_flap/link_flap_utils.py. It is not a good practice
for a common library to be dependent on function defined for specific
tests. Since the dependent function is just to show some status of DUT
while shuting down interfaces, the fix is to directly show the status of
DUT in library code instead of calling a function.

The third issue is that the library checks interfaces every 1 second after
shutdown or startup operation. This is a too aggressive interval. The fix is to
relax checking interval to 5 seconds.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>