Re-organize the documents
Currently there are many documents scattered in many folders without a
good structure. It is difficult to both read and to update documents.
This change re-organized location of the documents to have a better
structure and updated the reference links. Content of many documents
are still outdated. This PR is the first initiative of improving
sonic-mgmt documents.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>