Change IPv6 route prefix length from 96 to 64 in BGP BBR test
The current BGP BBR uses an IPv6 route with prefix length 96 in test. However, this rule could be blocked by
a route_map added by other features. The route_map rule only allows IPv6 routes with prefix length 64 from
T0 VMs. This change updated the IPv6 route prefix length from 96 to 64.

Another problem of this test is that the BGP BBR state is always restored to "enabled" in teardown. The fix
is to always restore BGP BBR state to the default_state defined in /etc/sonic/constants.py.

This PR also added checking of BGP BBR feature availability. If the feature is not supported, skip the test.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>