[qos] Fix param overwrite issue (#2616)
Signed-off-by: Neetha John <nejo@microsoft.com>

Some tests like headroom pool size have src port ids and dst port id hardcoded in qos.yml to ensure that we can actually fill up the headroom pool. In all the tests, we initialize the 'testParams' dict with values from qos.yml and then update it with a set of basicParams that we have derived. The basicParams includes src and dst port ids which will overwrite the ones that we pulled from qos.yml

How did you do it?
Consume the basicParams first and then update with values from qos.yml