Add test server side mux Y cable simulator
In dualtor testbed, mux Y cable is simulated by OVS in test server. An OVS
bridge is created for each of the mux. The PTF interface and test server
VLAN interfaces are attached to the bridge. The VLAN interfaces are
connected to DUT ports through fanout switches.
                          +--------------+
                          |              +----- upper_if
          PTF (host_if) --+  OVS bridge  |
                          |              +----- lower_if
                          +--------------+
Open flow rules are configured for the OVS bridge to simulate upstream
broadcasting and downstream dropping traffic from standby side.

To further simulate mux Y cable active/standby querying and setting, a
process need to be started in the test server. The process needs to expose
APIs for querying and setting active/standby status. On DUT side, a plugin
can be injected to intercept the calls to mux Y cable. Instead of calling
the actual mux driver functions, the APIs of mux simulator are called.
Then the process checks and updates open flow configurations of the OVS
bridge accordingly

This change added a program that can expose such HTTP APIs for the injected
plugin on DUT side to query and set mux Y cable active/standby status.

The program is a python script based on Flask. It is copied to test server
and started as background process by running 'testbed-cli.sh add-topo'
for dualtor type topology. It is stopped by running
'testbed-cli.sh remove-topo'.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>