Change IPv6 route prefix length from 96 to 64 in BGP BBR test (#2607)
What is the motivation for this PR?
The current BGP BBR uses an IPv6 route with prefix length 96 in test. However, this rule could be blocked by
a route_map added by other features. The route_map rule only allows IPv6 routes with prefix length 64 from
T0 VMs. Test failed because expected IPv6 routes are not found on DUT.

Another problem of this test is that the BGP BBR state is always restored to "enabled" in teardown. However,
the default_state of BBR is disabled in current image.

How did you do it?
Fixes:
* This change updated the IPv6 route prefix length from 96 to 64.
* Always restore BGP BBR state to the default_state defined in /etc/sonic/constants.py after test.
* Added checking of BGP BBR feature availability. If the feature is not supported, skip the test.

How did you verify/test it?
Test run the case on both physical and vs setup.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>