Fix test_fdb failure caused by FDB aging.
The test cast test_fdb is failing consistently on some testbed. The
root cause for the failure is the double loop among all interfaces costs
too much time, which is beyond the fdb age (10 minutes in default). This
PR address the issue by introducing a module level fixture to disable
FDB aging.

Signed-off-by: bingwang <bingwang@microsoft.com>