Adding test case to test telemetry service "osversion/build" query. (#2578)
Adding a new test case to validate the osversion/build telemetry query. This is a new query to get the Sonic build version from the telemetry service. Details of the changes on telemetry service can be found in this PR: Azure/sonic-telemetry#50

What is the motivation for this PR?
A test case needed to cover a recently added telemetry service feature.

How did you do it?
The test case runs the mentioned query and checks the existence of expected keywords in the response.

How did you verify/test it?
Using Sonic KVM testbed.

Co-authored-by: Murat <muacikgo@microsoft.com>