[test_fdb] Fix test_fdb failure caused by FDB aging. (#2658)
Description of PR
Summary:
Fixes # (issue)
The test case test_fdb is failing consistently on some testbed (like A7260). The root cause for the failure is the double loop among all interfaces costs too much time, which is beyond the fdb age (10 minutes in default). This PR address the issue by introducing a module level fixture to disable FDB aging.
This commit also clears the dataplane of testutil to ensure that the packet queue inside testutil is not filled up.

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 Test case(new/improvement)
Approach
What is the motivation for this PR?
This PR is to fix test_fdb failure caused by expired FDB entry.

How did you do it?
This PR address the issue by introducing a module level fixture to disable FDB aging.