[dualtor] Implement y_cable_simulator client (#2652)
Description of PR
Summary:
Fixes # (issue)
This PR is to introduce a simulator for y_cable for testing.
The simulator driver is injected into pmon container in deploy_mg for dualtor topo and hook the physical y_cable driver. The simulator interact with simulator driver running on vmhost server, and control the behavior of y_cable with open-flow.

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 Test case(new/improvement)
Approach
What is the motivation for this PR?
This commit contains the code for y_cable_simulator client.

How did you do it?
The simulator driver is injected to pmon container in 'deploy_mg' for dualtor topo and hook the physical y_cable driver. The simulator interact with simulator driver running on vmhost server, and control the behavior of y_cable with open-flow.