Try to disable apparmor for ntpd during add-topo
This is to address the NTP testing failed issue reported here:
https://github.com/Azure/sonic-buildimage/issues/5975
The reason is that apparmor is enabled for ntpd. The workaround
is to always try to disable apparmor for ntpd during add-topo.
If disabling failed, generally it is because apparmor is not
enabled for NTP on the test server. So, 'ignore_errors: yes' is
added to the tasks.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>