Add utilities to get ansible variables by hostname or group name (#2687)
This change added utilities to get ansible variables by hostname or group name.

What is the motivation for this PR?
In test code and library code, there is a frequent need of getting value of specific ansible variables belong to or is visible to specific ansible host or ansible group. This changed added utility functions to simplify this effort. By using these functions, we just need to supply the list of inventory files, server or group name, variable name.

How did you do it?
The functions will take advantage of the ansible's InventoryManager and VariableManager to get value of variables. Added functions:
* get_inventory_manager(inv_files)
* get_variable_manager(inv_files)
* get_host_vars(inv_files, hostname, variable=None)
* get_host_visible_vars(inv_files, hostname, variable=None)
* get_group_visible_vars(inv_files, group_name, variable=None)
* get_test_server_vars(inv_files, server, variable=None)
The inv_files argument can be a string or list. It's value should be inventory file path or list of inventory file paths. In tests, it can be retrieved from the pytest builtin fixture request, for example: request.config.getoption("ansible_inventory")

How did you verify/test it?
Import the utility file and test run the functions.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>