[mux] Implement packet drop interfaces for y_cable simulator client (#2675)
Description of PR
Summary:
Fixes # (issue)
This PR is to introduce new interfaces for packet drop in y_cable simulator. The interfaces interact with mux simulator server introduced in PR #2673 and control the flow on each interface.

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 Test case(new/improvement)
Approach
What is the motivation for this PR?
This PR is to introduce new interfaces for packet drop in y_cable simulator.

How did you do it?
Send HTTP requests to interact with mux simulator server introduced in PR #2673 and control the flow on each interface.

How did you verify/test it?
Verified on a dualtor-56 topo.

Verify all interfaces work as expected
Verify packets are dropped as expected
Verify no packet is dropped after recover.