Support for gen-mg to work with multi-dut where VM's connect to only a single host; support for supervisor card.
gen-mg supported multi-dut, but in a dualtor topology, each VM is assumed to
be connected to all the DUT's. To support the VM to be connected to a single DUT:
 - In vm_topo_config, the fields like interface_indexes, intfs, ip_intf etc would be
   either empty or null. Add checks for these where they are being used in gen-mg

gen-mg also assumes that all interfaces connected to a PortChannel are sequential. To fix this:
 - In config_sonic_based_on_testbed.yml, changed interface_to_vms structure to be compatable to
   allow using with_subelements to get the intf_names

For 'supervisor' card, there are no BGP peers, interfaces etc. required.
 - A 'supervisor' card is identified by having 'type' field in the inventory with value 'supervisor'.
 - Add this check to skip generation of portions not valid for supervisor card