[drop counters] Change test_egress_drop_on_down_link test logic (#2664)
What is the motivation for this PR?
The test_egress_drop_on_down_link test case was not passing. After investigation, I figured out that this is because when the test is shutting the neighbor link the ARP entry for that interface is cleared, so the packets destined for the neighbor IP hit the default route, and the packets are not dropped. Removing ARP entry after interface going down is normal behavior.

How did you do it?
Change test case logic and check that packets are not dropped.

Signed-off-by: Antonina Melnyk antoninax.melnyk@intel.com