Add caching capability for frequently used facts
To run test scripts, we frequently need to gather facts from various devices again and again. Most of the facts gatherings need to run some commands on remote devices through SSH connection, parse the commands output and return the results. Most of the time, the facts to be gathered are unchanged, like DUT HWSKU, platform, etc. For the less frequently changed facts, we can cache them for quicker access to save a lot of overhead for gathering them each time. Then we can improve the overall time required for running all the tests.

This change added caching capability for frequently used facts. For more details, please refer to the documentation included in this change.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>