Optimizing dut selection fixtures in pytest_generate_tests
This is a fix for issue #2790 - pytest_generate_tests in tests/conftest.py takes much more time than before

The reason for the issue was that creating the TestbedInfo takes 1-9 seconds, and was part of
pytest_generate_tests. Since pytest_generate_tests is called for all the tests/fixtures, this was
adding a long time when trying to execute many tests.

For some of the other dut selection fixtures we were also getting variables for the duts
from the inventory files - which also takes a long time - adding to the execution delay.

To fix is to use create the TestbedInfo and getting the variables from the inventory files
for all the DUTS only once - by storing it in global variables. If these global variables
are not set, then we would set them. Thus, creating TestbedInfo and getting variables only once
per pytest run