Make test_fdb dualtor ready (#2705)
Description of PR
Summary:
Fixes # (issue)
This PR is to fix test_fdb exception on dualtor testbed.
An exception will be thrown when test_fdb is running on dualtor testbed.
It's because the number of ports on ptf containes is larger than that on DUT, since there will be an injected interface for each DUT's interface connected to VM.
For example, there are 32 ports on a dualtor DUT, but there are 36 ports on ptf. So an exception will be thrown when attempting to map ptf port indices to DUT port indices.
This PR addressed the issue by adding a check for existence of idx in port_index_to_name.
What is the motivation for this PR?
This PR is to fix test_fdb on dualtor testbed.
How did you do it?
Adding a check for existence of idx in port_index_to_name.
How did you verify/test it?
Verified on both dualtor and single tor testbed.