Improve cache and enable caching of TestbedInfo
The original cache implementation uses json to save cached information to file. For objects that
cannot be serialized to JSON, they cannot be cached. Another limitation is that the cached facts
must be stored under sub-folder specified by hostname. This limited usage of the caching.

This change improved caching in these ways:
* Replace JSON with pickle to dump and load cached information. Then all object types supported by pickle can be cached.
* Use more generic 'zone' to specify sub-folder for cached facts. The zone can be hostname or other string.
* When 'testbed-cli.sh deploy-mg' is executed, cleanup all the cached information, not just cached information of current DUT.
* Enable caching of TestbedInfo which is also frequently used and less frequently changed.
* Updated related documents.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>