Refactor sanity check to use dynamic check fixtures (#2883)
What is the motivation for this PR?
The check items are defined in the checks.py module. In the original design, check item is defined as an ordinary function. All the dependent fixtures must be specified in the argument list of sanity_check. Then objects of the fixtures are passed to the check functions as arguments. However, this design has a limitation. Not all the sanity check dependent fixtures are supported on all topologies. On some topologies, sanity check may fail with getting those fixtures.

How did you do it?
To resolve that issue, this PR changed the design. Now the check items must be defined as fixtures. Then the check fixtures can be dynamically attached to test cases during run time. In the sanity check plugin, we can check the current testbed type or other conditions to decide whether or not to load certain check fixtures.

How did you verify/test it?
Use a customized test script to verify the combinations of sanity check supported marks and command line options.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>