[cache] ignore cache attribute error (#2892)
Approach
What is the motivation for this PR?
When building variable cache the first time, there will be an attribute error.

How did you do it?
Catch and ignore the initial cache error.

Signed-off-by: Ying Xie ying.xie@microsoft.com

How did you verify/test it?
Run test against a DUT doesn't have cache built (or remove the cache). Without the fix, there would be an exception. With the fix, the test proceeds.