Using cache instead of global variables in dut selection. Added 'enum_rand_one_per_hwsku_hostname' fixture.
With caching of facts, the tbInfo and vars in inventory for a host are cached. Instead of using
global variables '_dut_in_testbeds', _dut_vars, we would call the methods again, but since it is
cached, it is not re-creating testbed info or reading the vars from the inventory.

Added enum_rand_one_per_hwsku_hostname fixture to pytest_generate_tests. This is needed
as there are tests like platform tests (test_show_platform_syseeprom) that need to run
on both the supervisor and linecard of a T2 chassis.