[multi-asic] Update interface_facts to handle namespaces for multi_asic platforms (#2726)
Description of PR
Update needed to interface_facts anisible module to fetch interface details from namespaces in multi-asic platforms.

Approach
The approach taken is to carve out the logic of interface details retrieval into a python script named "/tmp/gather_intf_ip_info.py". This python script will be copied on to the DUT when this ansible module is run. This script will have the same functional logic as earlier to retrieve the interface details. This script will be invoked per namespace in case of multi-asic platforms and once per host in case of a single asic platform. The data collected is given back in the similar format as before.

The ansible module also accepts namespace as the parameter, where the interface facts from a specific namespace could be fetched.