[dualtor] Set selected ToR as active. (#2850)
Summary:
Fixes #2834
Current test framework selects a ToR randomly at the beginning of module. This PR introduces a autouse fixture to set the selected ToR as active.

Approach
What is the motivation for this PR?
This PR is to set randomly selected ToR as active.

How did you do it?
Add a module level autoused fixture to do this.