[multi-asic] Update drop counters tests for multi-asic platforms (#2730)
Update the drop counters tests to support multi-asic
Dependent on changes done in #2728 for common utilities change and #2726 for interface_facts change

How did you do it?

Introduced the interfaces_per_namespace field to populate the interfaces per namespace on test start.
find and pass asic_index to the tests. The asic_index is obtained from the port which is which is select randomly for test.

Add the namespace PREFIX/SUFFIX to certain commands if needed for l2/l3 interface stat/clear commands. There are certain commands which supports the "-namespace" option. The other commands are invoked with the "sudo ip netns exec " format. 

Use sonic-db-cli instead of redis-cli, as sonic-db-cli supports the namespace parameter.