Implement traffic distribution utility.
This PR implements a utility for checking balance in IPinIP tunnel. A
script will be running on ptf and generate traffic to standby ToR. The
packets should be forwarded to active ToR through IPinIP tunnel, and the
script on ptf will capture and verify if the traffic distribution is
balanced.

Signed-off-by: bingwang <bingwang@microsoft.com>