[device map] fix device map creation issue and caching issue
Device map creation could be aborted in the middle and leave
map half updated if an extra interface was not defined in the
topology but show up in the minigraph.

And because we cache the host variables on the file system,
the cache could prevent a correct map being generated.

Add a new pretest to remove cache. It would still be good to
generate cache for each nightly test.

Update kvmtest.sh to stop removing cache since the pretest
will do it now.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>