[wr_arp] add timeout to thread join (#2986)
What is the motivation for this PR?
test_wr_arp.py got stuck in nightly test until the test was aborted. This is because paramiko didn't return shell command return code from exec_command() call. So if there were error and warm-reboot didn't proceed, the command will return as success. Without return code, we cannot solely rely on stderr to judge if the command failed or not (because we could ignored the error and proceeded).

How did you do it?
When warm reboot fails to reboot the dut and test failed to detect the failure. The thread.join() would fail and the test will stuck
indefinitely.

Add a timeout to join and check if the thread quit after join returns.

Signed-off-by: Ying Xie ying.xie@microsoft.com

How did you verify/test it?
While warm reboot fails, with the fix, the thread join will timeout and give properly error information: