[dual tor]: Move y cable simulator driver injection to pretest (#2981)
Approach
What is the motivation for this PR?
With manual image changes, the y_cable_simulator_client on dual ToR testbeds is overwritten. Currently, a new minigraph needs to be deployed to inject the client again, which has a lot of overhead and is not desirable.

How did you do it?
Create a test case in test_pretest.py which creates and injects the simulator client to ToRs in a dual ToR topology.

How did you verify/test it?
./run_tests.sh -n <testbed_name> -i <inventory files> -u -c test_pretest.py::test_inject_y_cable_simulator_client

Verify that the file gets placed into /usr/lib/python3/dist-packages inside the pmon container on both ToRs.

Deploy minigraph, ensure that it completes successfully.

Any platform specific information?